Refactor out methods into multiple steps trying to make each part of the mapping clearer
