+++ Matter_Engine/Modify/MapMaterial.cs
@@ -93,178 +93,6 @@ namespace BH.Engine.Matter
 
         /***************************************************/
 
-        [Description("Maps a set of materials to a set of provided transdiciplinary materials.\n" +
-                     "First atempts to match the name of the provided materials to the material maps.\n" +
-                     "If no name match is found, atempts to instead find a material with as many matching MaterialProperties (based on type and name) as possible.\n" +
-                     "If a unique match is found based on one of the above matching methods, all Properties from the transdiciplinary material is applied to the material to be matched.")]
-        [Input("materials", "The Materials to Modify, will be evaluated based on their name and properties.")]
-        [Input("materialMaps", "The Material maps to match to. Should generally have unique names. Names of material as well as material properties will be used to map to the materials to be modified.")]
-        [Input("prioritiseMap", "Controls if main material or map material should be prioritised when conflicting information is found on both in terms of Density and/or Properties. If true, map is prioritised, if false, main material is prioritised.")]
-        [Input("uniquePerNamespace", "If true, the method is checking for similarity of MaterialProperties on the materials and found matching material map based on namespace. If false, this check is instead done on exact type.")]
-        [Output("materials", "Materials with modified list of properties. Materials for which no unique match could be found are unaffected.")]
-        public static IEnumerable<Material> MapMaterial(this IEnumerable<Material> materials, IEnumerable<Material> materialMaps, bool prioritiseMap = true, bool uniquePerNamespace = true)
-        {
-            if (materials.IsNullOrEmpty())
-                return null;
-
-            if (materialMaps == null || !materialMaps.Any())
-            {
-                Base.Compute.RecordWarning($"No {nameof(materialMaps)} provied. Unmapped {nameof(Material)}s returned.");
-                return materials;
-            }
-
-            ILookup<string, Material> nameLookup = materialMaps.ToLookup(x => x.Name);
-
-            Dictionary<Tuple<Type, string>, List<Material>> propertyMaps = new Dictionary<Tuple<Type, string>, List<Material>>();
-
-            foreach (Material mat in materialMaps)
-            {
-                foreach (IMaterialProperties property in mat.Properties)
-                {
-                    Tuple<Type, string> key = new Tuple<Type, string>(property.GetType(), property.Name);
-                    if (propertyMaps.ContainsKey(key))
-                        propertyMaps[key].Add(mat);
-                    else
-                        propertyMaps[key] = new List<Material>() { mat };
-                }
-            }
-
-            List<Material> mappedMaterials = new List<Material>();
-
-            foreach (Material material in materials)
-            {
-                mappedMaterials.Add(material.MapMaterial(nameLookup, propertyMaps, prioritiseMap, uniquePerNamespace));
-            }
-            return mappedMaterials;
-        }
-
-        /***************************************************/
-
-        [Description("Maps a set of materials in the MaterialCompositions to a set of provided transdiciplinary materials.\n" +
-                     "First atempts to match the name of the provided materials to the transdiciplinary material maps.\n" +
-                     "If no name match is found, atempts to instead find a material with as many matching MaterialProperties (based on type and name) as possible.\n" +
-                     "If a unique match is found based on one of the above matching methods, all Properties from the transdiciplinary material is applied to the material to be matched.")]
-        [Input("materialCompositions", "The MaterialCompositions to Modify. Materials int he MaterialCompositions will be evaluated based on the name and properties.")]
-        [Input("materialMaps", "The Material maps to match to. Should generally have unique names. Names of material as well as material properties will be used to map to the materials to be modified.")]
-        [Input("prioritiseMap", "Controls if main material or map material should be prioritised when conflicting information is found on both in terms of Density and/or Properties. If true, map is prioritised, if false, main material is prioritised.")]
-        [Input("uniquePerNamespace", "If true, the method is checking for similarity of MaterialProperties on the materials and found matching material map based on namespace. If false, this check is instead done on exact type.")]
-        [Output("materialCompositions", "MaterialComposition with Materials with modified list of properties. Materials for which no unique match could be found are unaffected.")]
-        public static IEnumerable<MaterialComposition> MapMaterial(this IEnumerable<MaterialComposition> materialCompositions, IEnumerable<Material> materialMaps, bool prioritiseMap = true, bool uniquePerNamespace = true)
-        {
-            if (materialCompositions.IsNullOrEmpty())
-                return null;
-
-            if (materialMaps == null || !materialMaps.Any())
-            {
-                Base.Compute.RecordWarning($"No {nameof(materialMaps)} provied. Unmapped {nameof(MaterialComposition)}s returned.");
-                return materialCompositions;
-            }
-            IEnumerable<Material> allMaterials = materialCompositions.SelectMany(x => x.Materials).GroupBy(x => x.BHoM_Guid).Select(x => x.First());
-            Dictionary<Guid, Material> mappedMaterials = allMaterials.MapMaterial(materialMaps, prioritiseMap, uniquePerNamespace).ToDictionary(x => x.BHoM_Guid);
-            return materialCompositions.Select(x => new MaterialComposition(x.Materials.Select(mat => mappedMaterials[mat.BHoM_Guid]), x.Ratios)).ToList();
-        }
-
-        /***************************************************/
-
-        [Description("Maps a set of materials in the MaterialCompositions to a set of provided transdiciplinary materials.\n" +
-                     "First atempts to match the name of the provided materials to the transdiciplinary material maps.\n" +
-                     "If no name match is found, atempts to instead find a material with as many matching MaterialProperties (based on type and name) as possible.\n" +
-                     "If a unique match is found based on one of the above matching methods, all Properties from the transdiciplinary material is applied to the material to be matched.")]
-        [Input("volumetricMaterialTakeoffs", "The VolumetricMaterialTakeoff to Modify. Materials int he VolumetricMaterialTakeoff will be evaluated based on the name and properties.")]
-        [Input("materialMaps", "The Material maps to match to. Should generally have unique names. Names of material as well as material properties will be used to map to the materials to be modified.")]
-        [Input("prioritiseMap", "Controls if main material or map material should be prioritised when conflicting information is found on both in terms of Density and/or Properties. If true, map is prioritised, if false, main material is prioritised.")]
-        [Input("uniquePerNamespace", "If true, the method is checking for similarity of MaterialProperties on the materials and found matching material map based on namespace. If false, this check is instead done on exact type.")]
-        [Output("volumetricMaterialTakeoffs", "MaterialComposition with Materials with modified list of properties. Materials for which no unique match could be found are unaffected.")]
-        public static IEnumerable<VolumetricMaterialTakeoff> MapMaterial(this IEnumerable<VolumetricMaterialTakeoff> volumetricMaterialTakeoffs, IEnumerable<Material> materialMaps, bool prioritiseMap = true, bool uniquePerNamespace = true)
-        {
-            if (volumetricMaterialTakeoffs.IsNullOrEmpty())
-                return null;
-
-            if (materialMaps == null || !materialMaps.Any())
-            {
-                Base.Compute.RecordWarning($"No {nameof(materialMaps)} provied. Unmapped {nameof(VolumetricMaterialTakeoff)}s returned.");
-                return volumetricMaterialTakeoffs;
-            }
-
-            IEnumerable<Material> allMaterials = volumetricMaterialTakeoffs.SelectMany(x => x.Materials).GroupBy(x => x.BHoM_Guid).Select(x => x.First());
-            Dictionary<Guid, Material> mappedMaterials = allMaterials.MapMaterial(materialMaps, prioritiseMap, uniquePerNamespace).ToDictionary(x => x.BHoM_Guid);
-            return volumetricMaterialTakeoffs.Select(x => new VolumetricMaterialTakeoff(x.Materials.Select(mat => mappedMaterials[mat.BHoM_Guid]), x.Volumes)).ToList();
-        }
-
-        /***************************************************/
-        /**** Private Methods                           ****/
-        /***************************************************/
-
-        [Description("Method preforming the mapping work. Matches the Material first by name, secondly by name of properties using the provided lookup and dictionary.")]
-        private static Material MapMaterial(this Material material, ILookup<string, Material> nameLookup, Dictionary<Tuple<Type, string>, List<Material>> propertyMaps, bool prioritiseMap, bool uniquePerNamespace)
-        {
-            List<Material> matches = nameLookup[material.Name].ToList();    //Try match by name
-
-            if (matches.Count == 0) //If no name match found, try match by material proeprties - Type and name
-            {
-                foreach (IMaterialProperties property in material.Properties)
-                {
-                    Tuple<Type, string> key = new Tuple<Type, string>(property.GetType(), property.Name);
-
-                    List<Material> propMatches;
-                    if (propertyMaps.TryGetValue(key, out propMatches))
-                        matches.AddRange(propMatches);
-                }
-
-                matches = matches.GroupBy(x => x.Name).Select(x => x.First()).ToList();
-                matches = matches.GroupBy(x => x.MatchScore(material))  //Get and group by score
-                                 .Where(x => x.Key > 0)                 //Only keep materials with score > 0, as negative score indicates at least one mismatch
-                                 .OrderByDescending(x => x.Key)         //Order by score - Higher score first
-                                 .FirstOrDefault()?.ToList() ?? new List<Material>();
-            }
-
-            if (matches.Count == 1) //Exactly one best match. Success!
-            {
-                return material.MergeProperties(matches[0], prioritiseMap, uniquePerNamespace);
-            }
-            else if (matches.Count == 0)    //No matches, record warning
-            {
-                Base.Compute.RecordWarning($"No map found for material named {material.Name}");
-                return material;
-            }
-            else  //More than one match. Record warning
-            {
-                Base.Compute.RecordWarning($"Material named {material.Name} has ambiguous matches to {string.Join(", ", matches.Select(x => x.Name))}. No mapping preformed.\nEnsure that the transdiciplinaryMaterialMaps have unique names and that your Material has a name matching the transdiciplinary material you want to match.");
-                return material;
-            }
-        }
-
-        /***************************************************/
-
-
-        [Description("Matches two materials based on how well the properties align based on name. For each aligning property the score is increased by 1. If a single misaligned property is found a score of -1 is returned.")]
-        private static int MatchScore(this Material mat1, Material mat2)
-        {
-            //Method matches two materials based on their properties.
-            //For each proeprty of the same type and same name, the score is increased by 1.
-            //If a property of the same type but a different name is found the score is set to -1, as that means a mismatch of properties.
-            int score = 0;
-            ILookup<Type, IMaterialProperties> matLookup = mat2.Properties.ToLookup(x => x.GetType());
-            foreach (IMaterialProperties prop in mat1.Properties)
-            {
-                string name = prop.Name;
-                Type type = prop.GetType();
-
-                IEnumerable<IMaterialProperties> prop2 = matLookup[type];
-                if (prop2.Any())
-                {
-                    if (prop2.Any(x => x.Name == name))
-                        score++;
-                    else
-                        return -1;
-                }
-
-            }
-            return score;
-        }
-
-        /***************************************************/
-
     }
 }
